requirements.txt
colorama
cython
numba
numpy
opencv-python
pygame
pyzmq